requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
decorator==5.1.1
Django==4.1.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
knox==0.1.14
loguru==0.6.0
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.3.0
pycparser==2.21
pyhcl==0.4.4
pytz==2022.6
requests==2.28.1
shortuuid==1.0.10
sqlparse==0.4.3
urllib3==1.26.12
validators==0.20.0
django-paypal==0.2.5



